test(file): moved test validation to its own file
- moved to its own file for easier reading
